requirements.txt
fastapi>=0.95.2,<0.100.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-groq>=0.1.0
langchain-community>=0.0.10
langchain-huggingface>=0.0.10
sentence-transformers>=2.2.0
chromadb>=0.4.0
pypdf>=3.17.1
beautifulsoup4>=4.12.0
requests>=2.31.0
